Updated ZooKeeperServiceRegistryTest to assert when services are registered/deregistered.
This may cause this test to become flaky.  I'm not entirely sure.  I've noticed some
variances in the code coverage statistics that seem related to the register method in
cases where it is unable to actually register a service.  I didn't expect this to happen
in a test case because the Curator TestingServer attempts to wait until ZooKeeper is up
before returning control back to the calling program.  In either case by adding these
assertions we make the test case less accepting of these timing conditions hopefully
with the result of finding the underlying cause in the future and resolving it.