Integrated health checks into the ServicePool.
ServicePool instances will now monitor failures that throw a ServiceException
by marking the ServiceEndpoint that was used for the failed callback as potentially
bad.  At that point it will no longer allow that ServiceEndpoint to be used to
service user requests until a health check for that endpoint is executed and
passes.  Health checks for failed endpoints are scheduled to run immediately
after a failed request, and then after that periodically once per minute.  Health
checks will then continue to run every minute until they pass, or until the bad
ServiceEndpoint is no longer visible to the HostDiscovery that the ServicePool
is using.

This change also updates the calculator example to run against an actual ZooKeeper
instance and to perform calculations in an infinite loop.  This provides a good
demo that can show services transitioning between being healthy and unhealthy
and going up and down.