Major refactorings to the service registry.
- The service registry has been refactored to have a helper class which is responsible for all of
the interaction with ZooKeeper.  The helper class is ZooKeeperPersistentEphemeralNode.

- ZooKeeperPersistentEphemeralNode adds a new primitive which is an ephemeral node along with the
maintenance code to ensure that as long as the node is not closed that a corresponding entry exists
in ZooKeeper.  It should handle all of the ZooKeeper ugliness.

- Added a new primitive to ZooKeeperTest as well.  It's a trigger which encapsulates a latch into
a Watcher.  These can be passed into checkExists calls, etc. during test cases and makes it so
only the trigger needs to be created, not a Watcher and a CountDownLatch.

- Upgraded Curator to version 1.1.12 which is compatible with ZooKeeper 3.4.x.