Refactored ServiceEndPoint as part of #4 (points 2 and 3).
This separates the concerns that were previously in SeviceEndpoint.  Specifically an endpoint
no longer knows anything about registration time or how to generate or parse JSON.  It also
no longer embeds registration time as that's a concern of the service registry.  In addition
ServiceEndpoint was renamed to ServiceEndPoint to be more standard.