Make ServicePool able to handle health checks that throw exceptions.
Previously we were susceptible to health checks that threw exceptions inadvertently
killing our background scheduled job to periodically monitor the health state of
a bad endpoint.  This is because the contract of a ScheduledExecutorService is that
if the runnable on a scheduled task throws an exception it doesn't reschedule the
task to run again in the future.  This change makes sure that we never let an
exception propagate out of a health check.