Simlified ServicePool.
This change simplifies two aspects of the service pool.  First it removes the
executeAsync method from the pool, instead implementing it in a separate class
AsyncServicePool that wraps a synchronous ServicePool.  Secondly, it simplifies
the ServicePoolTest by moving logical groups of tests out into their own test
classes: ServicePoolProxyTest, AsyncServicePoolTest, and the original
ServicePoolTest.

What this change doesn't do, is wire the AsyncServicePool into the
ServicePoolBuilder class.  So currently users don't have a dead simple easy
way to create them.  Instead they'll have to manually construct one via a
constructor call.  This can get fixed in a separate commit once the
details of how to do that are figured out.